rename: ErrorsCollection to NonEmptyErrorsChain, add scaladoc, cleanup API
